backend/requirements.txt
# ===== Core =====
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# ===== Database =====
sqlalchemy==2.0.36
alembic==1.14.1
aiosqlite==0.20.0

# ===== Data Collection =====
yfinance==0.2.51
vnstock==3.1.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# ===== Data Processing =====
pandas==2.2.3
numpy==2.2.1

# ===== Technical Analysis =====
ta==0.11.0

# ===== Machine Learning =====
scikit-learn==1.6.1
xgboost==2.1.3
lightgbm==4.5.0
tensorflow==2.18.0
prophet==1.1.6
arch==7.2.0

# ===== AI Reasoning =====
google-generativeai==0.8.4

# ===== Scheduling =====
apscheduler==3.10.4

# ===== Utilities =====
httpx==0.28.1
loguru==0.7.3
